pyproject.toml
[build-system]
requires = ["setuptools>=46.1.0", "setuptools_scm[toml]>=5"]
build-backend = "setuptools.build_meta"

[project]
name = "KnowMat"
dynamic = ["version"]
description = "AI agentic pipeline for extracting structured materials science data from PDFs and text."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Hasan Muhammad Sayeed", email = "hasan.sayeed.71.93@gmail.com" },
]
dependencies = [
  # LangChain & Orchestration
  "langchain>=0.3.0",
  "langchain-community>=0.3.0",
  "langchain-openai>=0.3.0",
  "langgraph>=0.2.0",
  "trustcall>=0.0.5",
  
  # OpenAI / API
  "openai>=1.0",
  
  # OCR & PDF parsing
  "beautifulsoup4>=4.12",
  "lxml>=6.0",
  "cssselect>=1.4",
  "einops>=0.8",
  "openpyxl>=3.1",
  "premailer>=3.10",
  "pymupdf>=1.24.0",
  "pdfminer.six>=20221105",
  
  # Validation and Configuration
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "python-dotenv>=1.0",
  "PyYAML>=6.0",
  "colorlog>=6.0",
  
  # Core utilities
  "tqdm>=4.66",
  "click>=8.0",
  "rich>=13.0",
  "pandas>=2.0",
  "pillow>=10.0",
  "markdownify>=0.13",
  
  # Composition parsing
  "pymatgen",
]

[project.optional-dependencies]
ocr = [
  "paddleocr>=3.0.0",
]
standardization = [
  "torch>=2.8.0",
  "torchvision>=0.23.0",
  "sentence-transformers>=5.0.0",
]
dev = [
  "pytest",
  "pytest-cov",
  "black",
  "isort",
  "ruff",
  "mypy",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*"]

[tool.setuptools.package-data]
knowmat = ["data/elements.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--cov", "knowmat", "--cov-report", "term-missing", "--verbose"]
norecursedirs = ["dist", "build", ".tox"]

[tool.isort]
profile = "black"
known_first_party = ["knowmat"]

[tool.coverage.run]
branch = true
source = ["knowmat"]

[tool.coverage.paths]
source = ["src/", "*/site-packages/"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self\\.debug",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    'if __name__ == "__main__":',
]

[tool.setuptools_scm]
version_scheme = "no-guess-dev"

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.ruff]
line-length = 100

